+++ roosevelt-less.js
@@ -1,6 +1,5 @@
 const fs = require('fs')
 const path = require('path')
-const LessPluginCleanCSS = require('less-plugin-clean-css')
 const less = require('less')
 
 module.exports = {
@@ -21,13 +20,7 @@ module.exports = {
 
       let lessInput = fs.readFileSync(path.join(app.get('cssPath'), fileName), 'utf8')
 
-      // Clean-css options
-      let opts
-
-      // LESS clean-css plugin
-      let cleanCSSPlugin
-
-      if (typeof params.sourceMap === 'object' && params.sourceMap !== null && params.sourceMap !== undefined && app.settings.env === 'development') {
+      if (typeof params.sourceMap === 'object' && app.settings.env === 'development') {
         options.sourceMap = params.sourceMap
       } else if (app.settings.env === 'development') {
         // Enable source mapping by default
@@ -37,20 +30,6 @@ module.exports = {
         }
       } else {
         options.sourceMap = undefined
-      }
-
-      // use clean-css plugin to minify CSS if minify param is true
-      if (app.get('params').minify) {
-        opts = params.cleanCSS || {}
-
-        // support use of the compress param
-        if (cssCompiler.compress) {
-          opts.advanced = true
-          opts.aggressiveMerging = true
-        }
-
-        cleanCSSPlugin = new LessPluginCleanCSS(opts)
-        options.plugins = [cleanCSSPlugin]
       }
 
       less.render(lessInput, options, (err, output) => {