+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next version
 
 - Enabled source mapping by default
+- Removed less-plugin-clean-css dependency and minification code
 
 ## 0.7.1 and below